Create Scala 2 veneer for Num.scala
